SUBMARINE-1207. Move files and change import path
