+++ submarine-server/server-api/src/main/java/org/apache/submarine/server/api/workbench/Role.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.submarine.server.workbench.entity;
+package org.apache.submarine.server.api.workbench;
 
 import java.util.List;
