+++ submarine-server/server-core/src/main/java/org/apache/submarine/server/workbench/rest/SysUserRestApi.java
@@ -23,10 +23,10 @@ import com.google.gson.Gson;
 import org.apache.submarine.server.workbench.annotation.SubmarineApi;
 import org.apache.submarine.server.workbench.database.entity.SysUserEntity;
 import org.apache.submarine.server.workbench.database.service.SysUserService;
-import org.apache.submarine.server.workbench.entity.Action;
-import org.apache.submarine.server.workbench.entity.Permission;
-import org.apache.submarine.server.workbench.entity.Role;
-import org.apache.submarine.server.workbench.entity.UserInfo;
+import org.apache.submarine.server.api.workbench.Action;
+import org.apache.submarine.server.api.workbench.Permission;
+import org.apache.submarine.server.api.workbench.Role;
+import org.apache.submarine.server.api.workbench.UserInfo;
 import org.apache.submarine.server.response.JsonResponse;
 import org.apache.submarine.server.response.JsonResponse.ListResult;
 import org.slf4j.Logger;
